properly crash starter too.

+++ release_tester/arangodb/starter/manager.py
@@ -327,9 +327,21 @@ class StarterManager():
 
     def crash_instances(self):
         """ make all managed instances plus the starter itself crash. """
+        try:
+            if (self.instance.status() == psutil.STATUS_RUNNING or
+                self.instance.status() == psutil.STATUS_SLEEPING):
+                print("generating coredump for " + str(self.instance))
+                psutil.Popen(['gcore', str(self.instance.pid)], cwd=self.basedir).wait()
+                
+                self.instance.send_signal(signal.SIGSEGV)
+                self.instance.wait()
+            else:
+                print("NOT generating coredump for " + str(self.instance))
+        except psutil.NoSuchProcess:
+            logging.info("instance already dead: " + str(self.instance))
+            
         for instance in self.all_instances:
             instance.crash_instance()
-        self.instance.send_signal(signal.SIGSEGV)
 
     def is_instance_running(self):
         """ check whether this is still running"""
